Factor out ir2gate-arithmetic.
