+++ xlsynth-g8r/src/gatify/mod.rs
@@ -3,5 +3,6 @@
 //! Gateification (PIR → `GateFn`) and associated pre-lowering rewrites.
 
 pub mod ir2gate;
+pub mod ir2gate_arithmetic;
 pub mod mul_by_const_csd;
 pub mod prep_for_gatify;
